image_service/requirements.txt
grpcio==1.80.0
grpcio-tools==1.80.0
pillow==12.2.0
protobuf==6.33.6
setuptools==82.0.1
typing_extensions==4.15.0